feat(redux): reorganizing and expanding the redux store to keep track of views
