+++ src/verdant/panel/events/event-date.tsx
@@ -1,13 +1,13 @@
 import * as React from "react";
 import { connect } from "react-redux";
-import { verdantState } from "../../redux/index";
 import { Checkpoint, CheckpointType } from "../../../lilgit/checkpoint";
 import {
+  verdantState,
   dateOpen,
   dateClose,
   saveBundles,
   eventState,
-} from "../../redux/events";
+} from "../../redux/";
 import NotebookEventDateBundle from "./event-date-bundle";
 
 /* CSS Constants */
