+++ src/verdant/redux/index.ts
@@ -1,121 +1,7 @@
-import { History } from "../../lilgit/history/";
-import { Nodey } from "../../lilgit/nodey/";
-import { Checkpoint } from "../../lilgit/checkpoint";
-import { ghostState, ghostReduce, ghostInitialState } from "./ghost";
-import { eventMapState, eventReducer, eventsInitialState } from "./events";
-import { Wishbone } from "../panel/details/wishbone";
-
-const SWITCH_TAB = "SWITCH_TAB";
-const FOCUS_CELL = "FOCUS_CELL";
-const INSPECT_TARGET = "INSPECT_TARGET";
-const SEARCH_FOR = "SEARCH_FOR";
-const INSPECT_ON = "INSPECT_ON";
-const INSPECT_OFF = "INSPECT_OFF";
-
-export const switchTab = (name: ActiveTab) => {
-  return {
-    type: SWITCH_TAB,
-    tab: name,
-  };
-};
-
-export const focusCell = (cell_index: number) => {
-  return {
-    type: FOCUS_CELL,
-    cell_index,
-  };
-};
-
-export const inspectNode = (target: Nodey) => {
-  return {
-    type: INSPECT_TARGET,
-    target,
-  };
-};
-
-export const inspectOn = () => {
-  return {
-    type: INSPECT_ON,
-  };
-};
-
-export const inspectOff = () => {
-  return {
-    type: INSPECT_OFF,
-  };
-};
-
-export const searchForText = (query: string) => {
-  return {
-    type: SEARCH_FOR,
-    query,
-  };
-};
-
-export enum ActiveTab {
-  Events,
-  Artifacts,
-  Artifact_Details,
-  Search,
-}
-
-export type artifactState = {
-  name: string;
-  ver: number;
-  outputVer?: number;
-  file?: string;
-};
-
-export type verdantState = {
-  getHistory: () => History;
-  currentEvent: Checkpoint;
-  activeTab: ActiveTab;
-  focusedCell: number;
-  inspectOn: boolean;
-  searchQuery: string;
-  inspectTarget: Nodey;
-  cellArtifacts: artifactState[];
-  notebookArtifact: artifactState;
-} & ghostState &
-  eventMapState;
-
-export const createInitialState = (getHistory: () => History): verdantState => {
-  return {
-    getHistory: getHistory,
-    currentEvent: null,
-    activeTab: ActiveTab.Events,
-    focusedCell: null,
-    inspectOn: false,
-    searchQuery: null,
-    inspectTarget: null,
-    cellArtifacts: [],
-    notebookArtifact: null,
-    ...eventsInitialState(),
-    ...ghostInitialState(),
-  };
-};
-
-export const verdantReducer = (state: verdantState, action: any) => {
-  switch (action.type) {
-    case SWITCH_TAB:
-      if (state.inspectOn) Wishbone.endWishbone(state.getHistory().notebook);
-      return { ...state, activeTab: action.tab, inspectOn: false };
-    case FOCUS_CELL:
-      return { ...state, focusedCell: action.cell_index };
-    case INSPECT_TARGET:
-      return { ...state, inspectTarget: action.target };
-    case INSPECT_ON:
-      if (!state.inspectOn) {
-        Wishbone.startWishbone(state.getHistory());
-      }
-      return { ...state, inspectOn: true };
-    case INSPECT_OFF:
-      if (state.inspectOn) Wishbone.endWishbone(state.getHistory().notebook);
-      return { ...state, inspectOn: false };
-    case SEARCH_FOR:
-      return { ...state, searchQuery: action.query };
-    default:
-      let state_ev = eventReducer(state, action);
-      return ghostReduce(state_ev, action);
-  }
-};
+export * from "./events";
+export * from "./ghost";
+export * from "./notebook";
+export * from "./state";
+export * from "./viewStates/artifact-detail";
+export * from "./viewStates/artifact-pane";
+export * from "./viewStates/search";
