+++ src/verdant/redux/viewStates/artifact-detail.ts
@@ -1,0 +1,9 @@
+export type artifactDetailState = {
+  openDetailPairs: string[];
+};
+
+export const artifactDetailInitialState = (): artifactDetailState => {
+  return {
+    openDetailPairs: [],
+  };
+};
