+++ src/verdant/verdant-notebook.ts
@@ -5,9 +5,8 @@ import { History } from "../lilgit/history/";
 import { VerNotebook } from "../lilgit/notebook";
 import { VerCell } from "../lilgit/cell";
 import { NotebookPanel } from "@jupyterlab/notebook";
-import { inspectNode, switchTab, ActiveTab } from "./redux/index";
+import { inspectNode, switchTab, ActiveTab, focusCell } from "./redux/";
 import { updateCheckpoint } from "./redux/events";
-import { focusCell } from "./redux";
 import { Nodey } from "../lilgit/nodey/";
 import { VerdantLog } from "./logger";
 import { NotebookEvent } from "../lilgit/notebook-events";
